update beta main.py

+++ main.py
@@ -123,15 +123,25 @@ def forwardPost(postURL=None, samePost=False, username=None):
 
 
 def follow(username):
+    errors = 0
     driver.get(f"https://www.instagram.com/{username}/")
     sleep(10)
     driver.find_element(
         By.XPATH, '//*[@id="react-root"]/section/main/div/ul/li[2]/a').click()
     sleep(5)
     for i in range(2, 16):
-        driver.find_element(
-            By.XPATH, f'//*[@id="react-root"]/section/main/div/ul/div/li[{i}]/div/div[2]/button').click()
-        sleep(2)
+        try:
+            driver.find_element(
+                By.XPATH, f'/html/body/div[1]/section/main/div/ul/div/li[{i}]/div/div[2]/button').click()
+            sleep(3)
+        except:
+            errors += 1
+            if errors >= 8:
+                break
+            if driver.find_elements(By.XPATH, '/html/body/div[4]/div/div/div/div[3]/button[2]'):
+                driver.find_element(
+                    By.XPATH, '/html/body/div[4]/div/div/div/div[3]/button[2]').click()
+                sleep(4)
 
 
 def unfollow(username):
@@ -363,57 +373,37 @@ while True:
             chrome_options.add_argument("--log-level=OFF")
             driver = webdriver.Chrome("chromedriver", options=chrome_options)
             # Create Image for post
-            if credentials.account[int(argv[1])][3] == 1:
-                with open(f"./userInputs/bgUser{argv[1]}.txt", "r+") as f:
-                    data = int(f.read())
-                    f.seek(0)
-                    f.write(str(data + 1))
-                    f.truncate()
-                    checkForCrashed = CreateImage("post",
-                                                  pickPost(), f'./CreateImage/{argv[1]}-{data%2}.png', data % 2)
-            else:
-                checkForCrashed = CreateImage("post",
-                                              pickPost(), f'./CreateImage/{argv[1]}.png')
-            # Create Image for story
-            # checkForCrashed = CreateImage("story",
-            #                               pickPost(), f'./CreateImage/s{argv[1]}.png')
-            # Upload the created image
-            login(int(argv[1]))
-            sendPost()
-            # Change To firefox for somereason
-            driver.quit()
-            # chrome_options = webdriver.ChromeOptions()
-            # chrome_options.add_argument("--auto-open-devtools-for-tabs")
-            # chrome_options.add_argument(
-            #     f"user-data-dir=./userInputs/Profile{argv[1]}/")
-            # chrome_options.add_argument("--no-sandbox")
-            # chrome_options.add_argument("--disable-dev-shm-usage")
-            # chrome_options.add_argument("--log-level=3")
-            # chrome_options.add_argument("--log-level=OFF")
-            # driver = webdriver.Chrome("chromedriver", options=chrome_options)
-            # Login
-            # login(int(argv[1]))
-            # Upload the created image
-            # sendStory()
+            if not posted:
+                if credentials.account[int(argv[1])][3] == 1:
+                    with open(f"./userInputs/bgUser{argv[1]}.txt", "r+") as f:
+                        data = int(f.read())
+                        f.seek(0)
+                        checkForCrashed = CreateImage(
+                            "post", pickPost(), f'./CreateImage/{argv[1]}-{data%2}.png', data % 2)
+                        if checkForCrashed != 'crash':
+                            f.write(str(data + 1))
+                        f.truncate()
+                else:
+                    checkForCrashed = CreateImage(
+                        "post", pickPost(), f'./CreateImage/{argv[1]}.png')
+                # Upload the created image
+                login(int(argv[1]))
+                sendPost()
+            posted = True
             # Follow few pages
-            # follow(
-            #     choice(credentials.followSource[credentials.account[int(argv[1])][2]]))
+            follow(
+                choice(credentials.followSource[credentials.account[int(argv[1])][2]]))
+            driver.quit()
+            sleep(5)
             # going for the next round
             runtimehour += 1
-            print(f"All done ! {runtimehour}/{TotalRunTime}")
             if runtimehour == TotalRunTime:
                 break
-            # driver.quit()
+            print(f"All done ! {runtimehour}/{TotalRunTime}")
             # here you can set the delay time
             sleep(choice(range(2260, 2530)))
+            posted = False
             reload(credentials)
-            driver = webdriver.Chrome("chromedriver", options=chrome_options)
         except Exception as excep:
             print(excep)
-            runtimehour += 1
-            if runtimehour == TotalRunTime:
-                break
             driver.quit()
-            sleep(choice(range(2260, 2530)))
-            reload(credentials)
-            driver = webdriver.Chrome("chromedriver", options=chrome_options)
